+++ app/src/renderer/components/staking/PageStaking.vue
@@ -18,14 +18,15 @@ tm-page(title='Staking')
     data-empty-search(v-else-if="!delegates.loading && sortedFilteredEnrichedDelegates.length === 0")
     template(v-else)
       panel-sort(:sort='sort', :properties="properties")
-      li-delegate(v-for='i in sortedFilteredEnrichedDelegates' :key='i.id' :delegate='i')
+      li-delegate(v-for='i in sortedFilteredEnrichedDelegates' :disabled="!userCanDelegate" :key='i.id' :delegate='i')
 
   .fixed-button-bar(v-if="!delegates.loading")
     template(v-if="userCanDelegate")
       .label #[strong {{ shoppingCart.length }}] validators selected
       tm-btn(id="go-to-bonding-btn" type="link" to="/staking/bond" :disabled="shoppingCart.length === 0" icon="chevron_right" icon-pos="right" value="Next" color="primary")
     template(v-else)
-      .label You do not have any {{bondingDenom}}s to stake.
+      .label(v-if="!delegation.loadedOnce && delegation.loading") Loading delegations...
+      .label(v-else) You do not have any {{bondingDenom}}s to stake.
       tm-btn(disabled icon="chevron_right" icon-pos="right" value="Next" color="primary")
 </template>